requirements.txt
nacl
requests
python-dotenv
